+++ GSDRoadSystem.cs
@@ -39,7 +39,7 @@ public class GSDRoadSystem : MonoBehaviour
         road.spline.road = road;
         road.GSDSplineObj = splineObj;
         road.GSDRS = this;
-        road.SetupUniqueIdentifier();
+        GSD.GSDRootUtil.SetupUniqueIdentifier(ref road.UID);
 
         road.ConstructRoad_ResetTerrainHistory();