+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored CheckLoadTexture into EditorUtilities
 - Removed unused editor timer
 - Unified DrawLine in EditorUtilities
+- Unified SetupUniqueIdentifier
 
 
 ## [Version: 1.9_FH] - 2019.04.22